+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ## 2.x.x (Released 2023-07-xx)
 
 - Added an ellipsis menu next to the build buttons for quick access to some of the build options.
+- Added a new validation check for notes that have vibrato, tremolo or slide, but have no sustain.
 - Fixed possible false positives for the "finger changes during slide check" when the anchor zone is wider than 4.
 
 ## 2.1.0 (Released 2023-07-24)